image-analysis/pyproject.toml
[project]
name = "image-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "pillow>=10.4",
    "piexif>=1.1",
    "tenacity>=9.0",
    "sqlalchemy>=2.0",
    "pydantic>=2.8",
    "openai>=1.60",
    "jsonschema>=4.23",
    "python-dateutil>=2.9",
    "numpy>=2.0",
]
